do not render post if no templates

+++ app.js
@@ -16,8 +16,7 @@ const { randomUUID } = require('node:crypto')
 
 const md = new MarkdownIt()
 if (createFolderPathifNotExists(path.join(ROOT, config.TEMPLATES_PATH))) {
-  console.log(`WARNING: ${config.TEMPLATES_PATH} directory does not exist. Created folder path but you have no templates. Ending process.`)
-  process.exit()
+  console.log(`WARNING: ${config.TEMPLATES_PATH} directory does not exist. Created folder path but you have no templates.`)
 }
 nunjucks.configure(path.join(ROOT, config.TEMPLATES_PATH), { autoescape: false })
 
@@ -27,11 +26,13 @@ const distFolderSetUp = () => {
   fs.rmSync(path.join(ROOT, config.OUTPUT_PATH), { recursive:true, force: true })
   createFolderPathifNotExists(path.join(ROOT, config.OUTPUT_PATH))
   createFolderPathifNotExists(path.join(ROOT, config.POST_OUTPUT_PATH))
+
   if (createFolderPathifNotExists(path.join(ROOT, config.STYLE_PATH))) {
     console.log(`WARNING: ${config.STYLE_PATH} does not exist. Created folder to compensate but you have no styles. Please restart app to see further style changes.`)
   } else {
     copyFilesFromPath(path.join(ROOT, config.STYLE_PATH), path.join(ROOT, config.OUTPUT_PATH))
   }
+  
   console.log("Successfuly created output folders and copied style sheets.")
 }
 
@@ -66,18 +67,24 @@ const generateBlogPosts = () => {
 
     // Render the page with new data and final HTML to output
     const BLOG_POST_NAME = config.BLOG_POST.replace(".html", '') // I don't trust users to name their templates properly
-    const output = nunjucks.render(`${BLOG_POST_NAME}.html`, {content: HTMLcontent, title: title, date: date.toDateString()})
-    fs.writeFileSync(path.join(ROOT, config.OUTPUT_PATH, '/posts/', `${filename}.html`), output)
+    if (fs.existsSync(path.join(ROOT, config.TEMPLATES_PATH, `${BLOG_POST_NAME}.html`))) {
+      const output = nunjucks.render(`${BLOG_POST_NAME}.html`, {content: HTMLcontent, title: title, date: date.toDateString()})
+      fs.writeFileSync(path.join(ROOT, config.OUTPUT_PATH, '/posts/', `${filename}.html`), output)
+    } else {
+      console.log(`ERROR: ROOT/${config.TEMPLATES_PATH}/${BLOG_POST_NAME}.html does not exist! Unable to generate post`)
+    }
   }
-  console.log('Successfully generated blog posts.')
 }
 
 const generateBlogIndex = postsData => {
   // Render the index page with all the posts
   const BLOG_INDEX_NAME = config.BLOG_INDEX.replace('.html', '') // I don't trust users
-  const result = nunjucks.render(`${BLOG_INDEX_NAME}.html`, {posts: postsData.reverse()})
-  fs.writeFileSync(path.join(ROOT, config.OUTPUT_PATH, `${BLOG_INDEX_NAME}.html`), result)
-  console.log('Successfully generated blog index.')
+  if (fs.existsSync(path.join(ROOT, config.TEMPLATES_PATH, `${BLOG_INDEX_NAME}.html`))) {
+    const result = nunjucks.render(`${BLOG_INDEX_NAME}.html`, {posts: postsData.reverse()})
+    fs.writeFileSync(path.join(ROOT, config.OUTPUT_PATH, `${BLOG_INDEX_NAME}.html`), result)
+  } else {
+    console.log(`ERROR: ROOT/${config.TEMPLATES_PATH}/${BLOG_INDEX_NAME}.html does not exist! Unable to generate post`)
+  }
 }
